Refactor exchange rate system with context, hooks, and new components
