+++ SMS_V.2.0/src/App.tsx
@@ -1,6 +1,7 @@
 import React, { Suspense, lazy } from 'react'
 import { BrowserRouter as Router, Routes, Route, Navigate } from 'react-router-dom'
 import { AuthProvider, useAuth } from './contexts/AuthContext'
+import { ExchangeRateProvider } from './contexts/ExchangeRateContext'
 import Header from './components/layout/Header'
 import Footer from './components/layout/Footer'
 import MusicPlayer from './components/layout/MusicPlayer'
@@ -127,8 +128,10 @@ function App() {
     <ErrorBoundary>
       <Router>
         <AuthProvider>
-          <PerformanceMonitor enableReporting={process.env.NODE_ENV === 'production'} />
-          <AppRoutes />
+          <ExchangeRateProvider>
+            <PerformanceMonitor enableReporting={process.env.NODE_ENV === 'production'} />
+            <AppRoutes />
+          </ExchangeRateProvider>
         </AuthProvider>
       </Router>
     </ErrorBoundary>
