+++ SMS_V.2.0/src/components/features/index.ts
@@ -1,0 +1,4 @@
+// Exchange Rate Components
+export { ExchangeRateModal } from './ExchangeRateModal';
+export { ExchangeRateDisplay } from './ExchangeRateDisplay';
+export { ExchangeRateExample } from './ExchangeRateExample';
