+++ SMS_V.2.0/src/contexts/ExchangeRateContext.tsx
@@ -1,0 +1,264 @@
+import React, { createContext, useContext, useEffect, useState, ReactNode } from 'react';
+import { supabase } from '../lib/supabase';
+import { useAuth } from './AuthContext';
+
+export interface ExchangeRate {
+  id?: string;
+  user_id: string;
+  usd_krw: number;
+  created_at?: string;
+  updated_at?: string;
+}
+
+interface ExchangeRateContextType {
+  rate: number;
+  lastUpdated: string;
+  isLoading: boolean;
+  error: string | null;
+  updateExchangeRate: (newRate: number) => Promise<boolean>;
+  updateWithLatestRate: () => Promise<boolean>;
+  convertCurrency: (amount: number, fromCurrency: string, toCurrency?: string) => number;
+  getFormattedRate: () => string;
+  getFormattedLastUpdated: () => string;
+  refetch: () => Promise<void>;
+}
+
+const ExchangeRateContext = createContext<ExchangeRateContextType | undefined>(undefined);
+
+const DEFAULT_RATE = 1300;
+
+interface ExchangeRateProviderProps {
+  children: ReactNode;
+}
+
+export const ExchangeRateProvider: React.FC<ExchangeRateProviderProps> = ({ children }) => {
+  const { user } = useAuth();
+  const [rate, setRate] = useState<number>(DEFAULT_RATE);
+  const [lastUpdated, setLastUpdated] = useState<string>('');
+  const [isLoading, setIsLoading] = useState<boolean>(true);
+  const [error, setError] = useState<string | null>(null);
+
+  // 환율 데이터 가져오기
+  const fetchExchangeRate = async () => {
+    if (!user) return;
+
+    try {
+      setIsLoading(true);
+      setError(null);
+
+      const { data, error: fetchError } = await supabase
+        .from('exchange_rates')
+        .select('*')
+        .eq('user_id', user.id)
+        .single();
+
+      if (fetchError && fetchError.code !== 'PGRST116') {
+        throw fetchError;
+      }
+
+      if (data) {
+        setRate(data.usd_krw);
+        setLastUpdated(data.updated_at || data.created_at || '');
+      } else {
+        // 기본 환율로 초기화
+        await createDefaultRate();
+      }
+    } catch (err) {
+      console.error('환율 데이터 가져오기 실패:', err);
+      setError('환율 데이터를 가져오는데 실패했습니다.');
+      setRate(DEFAULT_RATE);
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
+  // 기본 환율 생성
+  const createDefaultRate = async () => {
+    if (!user) return;
+
+    try {
+      const { data, error: createError } = await supabase
+        .from('exchange_rates')
+        .insert({
+          user_id: user.id,
+          usd_krw: DEFAULT_RATE
+        })
+        .select()
+        .single();
+
+      if (createError) throw createError;
+
+      setRate(data.usd_krw);
+      setLastUpdated(data.created_at || '');
+    } catch (err) {
+      console.error('기본 환율 생성 실패:', err);
+      setError('기본 환율을 생성하는데 실패했습니다.');
+    }
+  };
+
+  // 환율 업데이트
+  const updateExchangeRate = async (newRate: number): Promise<boolean> => {
+    if (!user) return false;
+
+    try {
+      setIsLoading(true);
+      setError(null);
+
+      // 입력 검증
+      if (newRate <= 0) {
+        throw new Error('환율은 0보다 커야 합니다.');
+      }
+
+      if (newRate > 10000) {
+        throw new Error('환율이 너무 높습니다.');
+      }
+
+      const { data, error: updateError } = await supabase
+        .from('exchange_rates')
+        .upsert({
+          user_id: user.id,
+          usd_krw: newRate,
+          updated_at: new Date().toISOString()
+        })
+        .select()
+        .single();
+
+      if (updateError) throw updateError;
+
+      setRate(data.usd_krw);
+      setLastUpdated(data.updated_at || data.created_at || '');
+      return true;
+    } catch (err) {
+      console.error('환율 업데이트 실패:', err);
+      setError(err instanceof Error ? err.message : '환율 업데이트에 실패했습니다.');
+      return false;
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
+  // 실시간 환율 API에서 최신 환율 가져오기
+  const fetchLatestRateFromAPI = async (): Promise<number | null> => {
+    try {
+      // 실제 API 키가 필요하므로 주석 처리
+      // const response = await fetch('https://api.exchangerate-api.com/v4/latest/USD');
+      // const data = await response.json();
+      // return data.rates.KRW;
+      
+      // 임시로 고정값 반환
+      return 1298;
+    } catch (err) {
+      console.error('실시간 환율 API 호출 실패:', err);
+      return null;
+    }
+  };
+
+  // 실시간 환율로 업데이트
+  const updateWithLatestRate = async (): Promise<boolean> => {
+    const latestRate = await fetchLatestRateFromAPI();
+    if (latestRate) {
+      return await updateExchangeRate(latestRate);
+    }
+    return false;
+  };
+
+  // 통화 변환 유틸리티
+  const convertCurrency = (amount: number, fromCurrency: string, toCurrency: string = 'KRW'): number => {
+    if (fromCurrency === toCurrency) return amount;
+    
+    if (fromCurrency === 'USD' && toCurrency === 'KRW') {
+      return amount * rate;
+    }
+    
+    if (fromCurrency === 'KRW' && toCurrency === 'USD') {
+      return amount / rate;
+    }
+    
+    return amount; // 지원하지 않는 통화 조합
+  };
+
+  // 포맷된 환율 문자열
+  const getFormattedRate = (): string => {
+    return `1 USD = ${rate.toLocaleString()} KRW`;
+  };
+
+  // 마지막 업데이트 시간 포맷팅
+  const getFormattedLastUpdated = (): string => {
+    if (!lastUpdated) return '업데이트 없음';
+    
+    const date = new Date(lastUpdated);
+    const now = new Date();
+    const diffInMinutes = Math.floor((now.getTime() - date.getTime()) / (1000 * 60));
+    
+    if (diffInMinutes < 1) return '방금 전';
+    if (diffInMinutes < 60) return `${diffInMinutes}분 전`;
+    if (diffInMinutes < 1440) return `${Math.floor(diffInMinutes / 60)}시간 전`;
+    
+    return date.toLocaleDateString('ko-KR', {
+      year: 'numeric',
+      month: 'short',
+      day: 'numeric',
+      hour: '2-digit',
+      minute: '2-digit'
+    });
+  };
+
+  // 초기 로드 및 실시간 구독
+  useEffect(() => {
+    if (user) {
+      fetchExchangeRate();
+
+      // 실시간 구독 설정
+      const subscription = supabase
+        .channel('exchange_rates_changes')
+        .on(
+          'postgres_changes',
+          {
+            event: '*',
+            schema: 'public',
+            table: 'exchange_rates',
+            filter: `user_id=eq.${user.id}`
+          },
+          (payload) => {
+            if (payload.eventType === 'UPDATE' || payload.eventType === 'INSERT') {
+              const newData = payload.new as ExchangeRate;
+              setRate(newData.usd_krw);
+              setLastUpdated(newData.updated_at || newData.created_at || '');
+            }
+          }
+        )
+        .subscribe();
+
+      return () => {
+        subscription.unsubscribe();
+      };
+    }
+  }, [user]);
+
+  const value: ExchangeRateContextType = {
+    rate,
+    lastUpdated,
+    isLoading,
+    error,
+    updateExchangeRate,
+    updateWithLatestRate,
+    convertCurrency,
+    getFormattedRate,
+    getFormattedLastUpdated,
+    refetch: fetchExchangeRate
+  };
+
+  return (
+    <ExchangeRateContext.Provider value={value}>
+      {children}
+    </ExchangeRateContext.Provider>
+  );
+};
+
+export const useExchangeRateContext = (): ExchangeRateContextType => {
+  const context = useContext(ExchangeRateContext);
+  if (context === undefined) {
+    throw new Error('useExchangeRateContext must be used within an ExchangeRateProvider');
+  }
+  return context;
+};
